Fix RFC/parser conformance bugs surfaced by round-1 + new TDD tests
Five bugs fixed, with new tests added that surface each one before
the corresponding production change:

1. EncodingSymbol::from_payload last-symbol length (RFC 5052 §9.1).
   The per-symbol emplace loop did not decrement remaining payload
   length, so the trailing partial symbol of a non-aligned source
   block reported len() == T instead of the residual. Receivers
   then wrote past the file buffer end on the last symbol of any
   file whose transfer_length was not a multiple of T. Round-1's
   GTEST_SKIP test is now active.

2. AlcPacket EXT_NOP/EXT_AUTH/EXT_TIME over-advance (RFC 5651 §3.2.5).
   The case body advanced hdr_ptr/consumed by a fixed +3 regardless
   of HEL, misaligning the next iteration's HET read. Fixed to
   advance by (HEL*4 - 2) bytes. Four new alc_packet_test cases
   cover HEL=1, HEL=2, two consecutive EXT_NOPs, and EXT_AUTH +
   EXT_TIME interleaved with EXT_FDT.

3. calculate_md5 broken error sentinel. The function returned -1 on
   invalid input but its return type was `unsigned int`, silently
   turning the sentinel into UINT_MAX. The File constructor's
   `if (calculate_md5(...) < 0)` check therefore never fired, so a
   File built from a length-0 buffer produced an FDT entry with an
   uninitialised MD5. Return type changed to int. New file_test.cpp
   covers (null,0), (non-null,0), and the valid-input path, plus a
   File construction test that now correctly throws.

4. AlcPacket EXT_FDT FLUTE-version check (RFC 6726 §3.4.1). Parser
   accepted only Version=1 (RFC 3926 / FLUTE v1). RFC 6726 mandates
   Version=2 for FLUTE v2 sessions, the basis of the MBMS profile.
   Now accepts both 1 and 2; anything else throws. Round-1's
   GTEST_SKIP test is now active, plus a new RejectsExtFdtVersionThree-
   AndAbove test for the negative case.

5. AlcPacket producer codepoint overwrite. Line 308 overwrote the
   carefully-set LCT codepoint with the raw FecScheme enum value.
   Latent today (CompactNoCode==0, Raptor==1 happen to match), but
   would assign wrong codepoints if any other FEC scheme were ever
   accepted. Removed.

ctest before: 39 active / 13 skipped. After: 50 active / 11 skipped
(round-1 skips for partial-last-symbol and FLUTE-v2 are now active;
the 11 remaining skips all document MBMS extension fields that the
FileEntry struct doesn't yet surface — out of scope for this round).
Verified gates: TESTING=ON, TESTING=OFF, ENABLE_RAPTOR=ON.

+++ src/File.cpp
@@ -295,9 +295,9 @@ auto LibFlute::File::mark_completed(const std::vector<EncodingSymbol>& symbols, 
   }
 }
 
-auto LibFlute::calculate_md5(char *input, size_t length, unsigned char *result) -> unsigned int
+auto LibFlute::calculate_md5(char *input, size_t length, unsigned char *result) -> int
 {
-  // simple implementation based on openssl docs (https://www.openssl.org/docs/man3.0/man3/EVP_DigestInit_ex.html) 
+  // simple implementation based on openssl docs (https://www.openssl.org/docs/man3.0/man3/EVP_DigestInit_ex.html)
   if (input == nullptr || length == 0U) {
     spdlog::error("MD5 called with invalid input");
     return -1;
@@ -319,5 +319,5 @@ auto LibFlute::calculate_md5(char *input, size_t length, unsigned char *result) 
   }
   spdlog::debug("MD5 Digest is {}", buf);
 
-  return md_len;
+  return static_cast<int>(md_len);
 }
